+++ src/ff/mod.rs
@@ -1,0 +1,7 @@
+// ff - Finite Fields
+//
+// This is where we store arithmetic shared secret data models.
+
+mod field;
+
+pub use field::{Field, Fp31};
