+++ src/lib.rs
@@ -3,6 +3,7 @@
 pub mod chunkscan;
 pub mod cli;
 pub mod error;
+pub mod ff;
 pub mod helpers;
 pub mod protocol;
 pub mod secret_sharing;
