+++ src/protocol/securemul.rs
@@ -1,7 +1,8 @@
 use crate::error::BoxError;
+use crate::ff::Field;
 use crate::helpers::{fabric::Network, Direction};
 use crate::protocol::{context::ProtocolContext, RecordId};
-use crate::secret_sharing::{Field, Replicated};
+use crate::secret_sharing::Replicated;
 use serde::{Deserialize, Serialize};
 use std::fmt::Debug;
 
@@ -70,141 +71,13 @@ impl<'a, N: Network, F: Field> SecureMul<'a, N, F> {
     }
 }
 
-#[derive(Error, Debug)]
-pub enum Error {}
-
-/// Module to support streaming interface for secure multiplication
-pub mod stream {
-    use crate::protocol::context::ProtocolContext;
-    use crate::secret_sharing::{Field, Replicated};
-    use futures::Stream;
-
-    use crate::chunkscan::ChunkScan;
-    use crate::helpers::fabric::Network;
-    use crate::protocol::{RecordId, Step};
-
-    #[derive(Copy, Clone, PartialEq, Eq, Debug, Hash)]
-    pub struct StreamingStep;
-
-    impl Step for StreamingStep {}
-    impl AsRef<str> for StreamingStep {
-        fn as_ref(&self) -> &str {
-            "streaming"
-        }
-    }
-
-    /// Consumes the input stream of replicated secret shares and produces a new stream with elements
-    /// being the product of items in the input stream. For example, if (a, b, c) are elements of the
-    /// input stream, output will contain two elements: (a*b, a*b*c)
-    ///
-    /// ## Panics
-    /// Panics if one of the internal invariants does not hold.
-    #[allow(dead_code)]
-    pub fn secure_multiply<'a, F, N, S>(
-        input_stream: S,
-        ctx: &'a ProtocolContext<'a, N>,
-        _index: u128,
-    ) -> impl Stream<Item = Replicated<F>> + 'a
-    where
-        S: Stream<Item = Replicated<F>> + 'a,
-        F: Field,
-        N: Network,
-    {
-        let mut stream_element_idx = 0;
-
-        // TODO (alex): is there a way to deal with async without pinning stream to the heap?
-        Box::pin(ChunkScan::new(
-            input_stream,
-            2, // buffer two elements
-            move |mut items: Vec<Replicated<F>>| async move {
-                debug_assert!(items.len() == 2);
-
-                let b_share = items.pop().unwrap();
-                let a_share = items.pop().unwrap();
-                stream_element_idx += 1; // TODO(mt): revisit (use enumerate()?)
-
-                ctx.multiply(RecordId::from(stream_element_idx))
-                    .await
-                    .execute(a_share, b_share)
-                    .await
-            },
-        ))
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use crate::helpers::Identity;
-        use crate::protocol::context::ProtocolContext;
-        use crate::protocol::securemul::stream::secure_multiply;
-        use crate::protocol::QueryId;
-        use crate::secret_sharing::{Fp31, Replicated};
-        use crate::test_fixture::{logging, make_world, share, validate_and_reconstruct};
-        use futures::StreamExt;
-        use futures_util::future::join_all;
-        use futures_util::stream;
-        use rand::rngs::mock::StepRng;
-
-        /// Secure multiplication may be used with Stream API where shares are provided as elements
-        /// of a `Stream`.
-        #[tokio::test]
-        async fn supports_stream_of_secret_shares() {
-            // beforeEach is not a thing in Rust yet: https://github.com/rust-lang/rfcs/issues/1664
-            logging::setup();
-
-            // we compute a*b*c in this test. 4*3*2 = 24
-            let mut rand = StepRng::new(1, 1);
-            let a = share(Fp31::from(4_u128), &mut rand);
-            let b = share(Fp31::from(3_u128), &mut rand);
-            let c = share(Fp31::from(2_u128), &mut rand);
-            let start_index = 1024_u128;
-
-            // setup helpers
-            let world = make_world(QueryId);
-
-            // dedicated streams for each helper
-            let input = [
-                stream::iter(vec![a[0], b[0], c[0]]),
-                stream::iter(vec![a[1], b[1], c[1]]),
-                stream::iter(vec![a[2], b[2], c[2]]),
-            ];
-
-            // create 3 tasks (1 per helper) that will execute secure multiplication
-            let handles = input
-                .into_iter()
-                .zip(world.participants)
-                .zip(world.gateways)
-                .zip([Identity::H1, Identity::H2, Identity::H3])
-                .map(|(((input, prss), gateway), role)| {
-                    tokio::spawn(async move {
-                        let ctx = ProtocolContext::new(role, &prss, &gateway);
-                        let mut stream = secure_multiply(input, &ctx, start_index);
-
-                        // compute a*b
-                        let _ = stream.next().await.expect("Failed to compute a*b");
-
-                        // compute (a*b)*c and return it
-                        stream.next().await.expect("Failed to compute a*b*c")
-                    })
-                });
-
-            let result_shares: [Replicated<Fp31>; 3] =
-                join_all(handles.map(|handle| async { handle.await.unwrap() }))
-                    .await
-                    .try_into()
-                    .unwrap();
-            let result_shares = (result_shares[0], result_shares[1], result_shares[2]);
-
-            assert_eq!(Fp31::from(24_u128), validate_and_reconstruct(result_shares));
-        }
-    }
-}
-
 #[cfg(test)]
 pub mod tests {
     use crate::error::BoxError;
+    use crate::ff::{Field, Fp31};
     use crate::helpers::fabric::Network;
     use crate::protocol::{context::ProtocolContext, QueryId, RecordId};
-    use crate::secret_sharing::{Field, Fp31, Replicated};
+    use crate::secret_sharing::Replicated;
     use crate::test_fixture::{
         fabric::InMemoryEndpoint, logging, make_contexts, make_world, share,
         validate_and_reconstruct, TestWorld,
