+++ src/secret_sharing/malicious_replicated.rs
@@ -6,8 +6,9 @@ use std::{
     ops::{Add, Mul, Neg, Sub},
 };
 
+use crate::ff::Field;
 use crate::helpers::Identity;
-use crate::secret_sharing::{Field, Replicated};
+use crate::secret_sharing::Replicated;
 
 #[derive(Clone, Copy, PartialEq, Eq)]
 pub struct MaliciousReplicated<F> {
@@ -109,8 +110,8 @@ impl<F: Field> Mul<F> for MaliciousReplicated<F> {
 #[cfg(test)]
 mod tests {
     use super::MaliciousReplicated;
+    use crate::ff::{Field, Fp31};
     use crate::helpers::Identity;
-    use crate::secret_sharing::{Field, Fp31};
     use crate::test_fixture::{share, validate_and_reconstruct};
     use proptest::prelude::Rng;
